move console commands to separate package
